Code now compiles
Trying to work out if there is any way to -- or point in -- buffering
the mapperCursor object. In particular I have been studying the
interaction between NextInterval() on the Mapper, and Next(), which is
called by the map functions.

There may not be, and perhaps I just copy what is there. Seek() is only
called once, so that could be a passthru. Next() on the cursor is
called, and the values checked. Depending on those values, the buffers
are set, but there not buffers of the values. They are more like flags
sometimes, with k=0 indicating the cursor is finished for this interval.